*** Don't force copy if exists

+++ openlegacy/openlegacy-providers/openlegacy-h3270/src/main/java/org/openlegacy/providers/h3270/H3270TerminalConnectionFactory.java
@@ -127,7 +127,9 @@ public class H3270TerminalConnectionFactory implements TerminalConnectionFactory
 	private void initResource(String resourceName) throws IOException {
 		File targetFile = new File(initWorkingDir(), resourceName);
 		targetFile.getParentFile().mkdirs();
-		IOUtils.copy(getClass().getResourceAsStream(resourceName), new FileOutputStream(targetFile));
+		if (!targetFile.exists()) {
+			IOUtils.copy(getClass().getResourceAsStream(resourceName), new FileOutputStream(targetFile));
+		}
 	}
 
 	private void initConfiguration(File targetFile) {
